Minicute.Parser.Common.Lexer

Description

Common lexer functions used in miniCUTE compiler

Documentation

betweenRoundBrackets :: MonadParser e s m => m a -> m a

betweenRoundBrackets p parses "(", and then p, and finally ")".

gMachineIdentifier :: MonadParser e s m => m (Tokens s)

gMachineIdentifier parses any character sequences not containing ';' or space characters as an identifier.

identifier :: MonadParser e s m => m (Tokens s)

identifier parses any character sequences starting with identifierFirstChar followed by identifierRestChar. For example, this function parses "_abc", "ab4", and "a_ef_3" successfully.

For more examples, see the test module for this function.

keyword :: MonadParser e s m => Tokens s -> m (Tokens s)

keyword parses any keywords.

The keywords are let, letrec, in, match, with, and $C.

symbol :: MonadParser e s m => Tokens s -> m ()

symbol str parses str and ignore the result.

integer :: (MonadParser e s m, Integral a) => m a

integer parses a decimal integer with no prefix, binary integer with "0b" or "0B" prefix, octal integer with "0o" or "0O" prefix, decimal integer with "0d" or "0D" prefix, hexadecimal integer with "0x" or "0X" prefix.

spacesConsumer :: MonadParser e s m => m ()

spacesConsumer consumes all consecutive spaces and ignore the result.